# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "vLLM CI Dashboard API",
    "version": "1.0.0",
    "description": "Public read-only API backing https://ci.vllm.ai. All documented GET endpoints are unauthenticated and return JSON; errors use `{ \"error\": string }`. Responses are cached via `Vercel-CDN-Cache-Control`; treat data as eventually consistent (30s-10min staleness)."
  },
  "servers": [
    {
      "url": "https://ci.vllm.ai",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/perf": {
      "get": {
        "operationId": "get-perf",
        "summary": "Benchmark performance data",
        "description": "",
        "tags": [
          "Perf"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "model",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "device",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "tp",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "conc",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "start",
            "schema": {
              "type": "string",
              "description": "ISO date (YYYY-MM-DD); omitted or invalid falls back to 2026-06-14"
            },
            "required": false,
            "description": "ISO date (YYYY-MM-DD); omitted or invalid falls back to 2026-06-14"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/perf/filters": {
      "get": {
        "operationId": "get-perf-filters",
        "summary": "Available perf filter values",
        "description": "Distinct models (with per-model datapoint counts), devices, tps, concs, precisions, and images in the given date range. Note: tps/concs are returned as strings even though `/api/perf` accepts them as integers.",
        "tags": [
          "Perf"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "schema": {
              "type": "string",
              "description": "ISO date (YYYY-MM-DD); omitted or invalid falls back to 2026-06-14"
            },
            "required": false,
            "description": "ISO date (YYYY-MM-DD); omitted or invalid falls back to 2026-06-14"
          },
          {
            "in": "query",
            "name": "end",
            "schema": {
              "type": "string",
              "description": "ISO date (YYYY-MM-DD); ignored if not a valid date"
            },
            "required": false,
            "description": "ISO date (YYYY-MM-DD); ignored if not a valid date"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/queue": {
      "get": {
        "operationId": "get-queue",
        "summary": "Historical queue wait-time statistics",
        "description": "Bucket width adapts to the requested span (5 minutes up to 1 week).",
        "tags": [
          "Queue"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "pipeline",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "queue",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
...
}